+++ intelmq/lib/message.py
@@ -49,10 +49,10 @@ class MessageFactory:
             MessageFactory.unserialize
             MessageFactory.serialize
         """
-        if default_type and "__type" not in message:
-            message["__type"] = default_type
+        if not default_type and '__type' not in message:
+            raise ValueError("Message type could not be determined. Input message misses '__type' and parameter 'default_type' not given.")
         try:
-            class_reference = getattr(intelmq.lib.message, message["__type"])
+            class_reference = getattr(intelmq.lib.message, message.get("__type", default_type))
         except AttributeError:
             raise exceptions.InvalidArgument('__type',
                                              got=message["__type"],
@@ -60,6 +60,8 @@ class MessageFactory:
                                              docs=HARMONIZATION_CONF_FILE)
         # don't modify the parameter
         message_copy = message.copy()
+        if default_type and "__type" not in message_copy:
+            message_copy["__type"] = default_type
         del message_copy["__type"]
         return class_reference(message_copy, auto=True, harmonization=harmonization)