+++ intelmq/tests/bots/parsers/json/test_parser.py
@@ -6,6 +6,7 @@
 import base64
 import os
 import unittest
+from json import loads as json_loads, dumps as json_dumps
 
 import intelmq.lib.test as test
 from intelmq.bots.parsers.json.parser import JSONParserBot
@@ -51,6 +52,21 @@ NO_DEFAULT_REPORT = {"feed.name": "Test feed",
 NO_DEFAULT_EVENT = MULTILINE_EVENTS[1].copy()
 NO_DEFAULT_EVENT['raw'] = base64.b64encode(b'{"source.ip": "127.0.0.2", "classification.type": "c2-server"}\n').decode()
 
+with open(os.path.join(os.path.dirname(__file__), 'ncscnl.json'), 'rb') as fh:
+    NCSCNL_FILE = fh.read()
+NCSCNL_RAW = base64.b64encode(NCSCNL_FILE).decode()
+NCSC_EVENTS = json_loads(NCSCNL_FILE)
+for i, event in enumerate(NCSC_EVENTS):
+    NCSC_EVENTS[i]['raw'] = base64.b64encode(json_dumps(event, sort_keys=True).encode()).decode()
+    NCSC_EVENTS[i]['classification.type'] = 'undetermined'
+    NCSC_EVENTS[i]['feed.name'] = 'NCSC.NL'
+    NCSC_EVENTS[i]['__type'] = 'Event'
+
+NCSCNL_REPORT = {"feed.name": "NCSC.NL",
+                 "raw": NCSCNL_RAW,
+                 "__type": "Report",
+                 }
+
 
 class TestJSONParserBot(test.BotTestCase, unittest.TestCase):
     """
@@ -70,8 +86,7 @@ class TestJSONParserBot(test.BotTestCase, unittest.TestCase):
     def test_multiline_report(self):
         """ Test if correct Event has been produced. """
         self.input_message = MULTILINE_REPORT
-        self.sysconfig = {"splitlines": True}
-        self.run_bot()
+        self.run_bot(parameters={"splitlines": True})
         self.assertMessageEqual(0, MULTILINE_EVENTS[0])
         self.assertMessageEqual(1, MULTILINE_EVENTS[1])
 
@@ -80,6 +95,14 @@ class TestJSONParserBot(test.BotTestCase, unittest.TestCase):
         self.input_message = NO_DEFAULT_REPORT
         self.run_bot()
         self.assertMessageEqual(0, NO_DEFAULT_EVENT)
+
+    def test_ncscnl(self):
+        """ A file containing a list of events (not per line) """
+        self.input_message = NCSCNL_REPORT
+        self.run_bot(parameters={'multiple_events': True})
+        self.assertMessageEqual(0, NCSC_EVENTS[0])
+        self.assertMessageEqual(1, NCSC_EVENTS[1])
+        self.assertMessageEqual(2, NCSC_EVENTS[2])
 
 
 if __name__ == '__main__':  # pragma: no cover